refactor(parser): move callbacks in separate file
