+++ source/source-full3.md
@@ -1,0 +1,38 @@
+#~ this is previewText
+~[Ship the data importer you always dreamed of][https://bit.ly/30AWyC4][https://raw.githubusercontent.com/atherdon/newsletters/master/archive/logos/thematic/2020/September/flatfile-logo-black-small-vertical.png]
+
+# Paid Video Call Opportunities for [Nominee Name]? (beta)
+
+** I AM SO BOLD and _ME TOO_ **    
+
+**_YEP_**
+   
+***YEP*** 
+
+Hey [writer name],
+
+Good news, top Hacker Noon writers can now list themselves for paid video calls via Hacker Noon’s new Meet Tech Experts Program (beta). As a Noonie nominee, I thought you may be interested in joining this program. 
+
+If interested in receiving leads for paid video calls, [please fill out this Google Form](https://docs.google.com/forms/d/e/1FAIpQLSce_si23Kc8Ydmv27J9Z4f_BdwY5VWczKDGtZBGtUPCsbDW0Q/viewform). We’ll review the information and then if sufficient send over an invite to the Hacker Noon network via Superpeer within a week. This program is a revenue share with 75% of the earnings going to the writer and 25% going to Hacker Noon - minus the Superpeer transaction fee. Writers set their own rate and availability for all video calls. 
+
+Meet Tech Experts by Hacker Noon is a beta marketplace. Would be great to list a talented contributor like yourself to start to stimulate supply side demand. 
+
+Kind Regards
+
+*David Smooke*
+
+*   Hacker Noon Stickers
+*   A standard .TECH domain for 3 years
+*   A standard .TECH domain for 2 years
+    *   Sub item 1
+    *   Sub item 2
+    *   Sub item 3
+*   A standard .TECH domain for 5 years
+
+
+## The Bottom Line
+
+
+### The Bottom Line
+
+~[Ship the data importer you always dreamed of][https://bit.ly/30AWyC4][https://raw.githubusercontent.com/atherdon/newsletters/master/archive/logos/thematic/2020/September/flatfile-logo-black-small-vertical.png]
